join 64 players

+++ src/game/server/eventhandler.h
@@ -3,7 +3,8 @@
 #ifndef GAME_SERVER_EVENTHANDLER_H
 #define GAME_SERVER_EVENTHANDLER_H
 
-//
+#include <base/system.h>
+
 class CEventHandler
 {
 	static const int MAX_EVENTS = 128;
@@ -12,7 +13,7 @@ class CEventHandler
 	int m_aTypes[MAX_EVENTS]; // TODO: remove some of these arrays
 	int m_aOffsets[MAX_EVENTS];
 	int m_aSizes[MAX_EVENTS];
-	int m_aClientMasks[MAX_EVENTS];
+	int64 m_aClientMasks[MAX_EVENTS];
 	char m_aData[MAX_DATASIZE];
 
 	class CGameContext *m_pGameServer;
@@ -24,7 +25,7 @@ public:
 	void SetGameServer(CGameContext *pGameServer);
 
 	CEventHandler();
-	void *Create(int Type, int Size, int Mask = -1);
+	void *Create(int Type, int Size, int64 Mask = -1LL);
 	void Clear();
 	void Snap(int SnappingClient);
 };
